backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.0

pydantic[email]==2.5.3
pytest==7.4.0
httpx==0.24.1

### Files ###
python-multipart==0.0.7

### DB ###
SQLAlchemy==2.0.13
psycopg2==2.9.9
alembic==1.11.1

### Redis ###
aiofiles==23.2.1
aioredis==2.0.1

### JWT Tokeen ### 
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0